+++ editor/editor-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-core
 
+## 150.0.1
+
+### Patch Changes
+
+- [`52aed763767`](https://bitbucket.org/atlassian/atlassian-frontend/commits/52aed763767) - [ED-13879] deletes a placeholder text node when opening a typeahead
+
 ## 150.0.0
 
 ### Major Changes